Add social media buttons and update user avatar in MainUserComponent
Added Github, Facebook, Twitter, and Website buttons in MainUserComponent.vue. Enhanced user avatar to display either the user's avatar if available, or a default, and updated social media buttons to only appear if a link is provided by the user. This should provide a more personalized experience for the user and facilitate social media integration.