Add visibility toggle for CardUpdateUser in store
A new visibility toggle for the CardUpdateUser component has been added in the Vuex store. This includes its definition in the application state, a dedicated mutation function, and an action for toggling visibility. These changes will facilitate state management for the CardUpdateUser component; at the same time, giving us better control and trackability.