Replace v-show with v-if in NavbarComponent
Changed the directive from v-show to v-if in the 'CandidateModalLogin' component within 'NavbarComponent.vue'. This switch sees to it that modalVisibility value is removed from the DOM when it's false, rather than just hidden. This can give a slight performance improvement in cases where the visibility of the modal changes infrequently.