Add avatar upload feature to MainUserComponent
Implemented an avatar upload feature to the MainUserComponent. Importing the 'uploadAvatar' method from FileManager and adding a file input and handling function in the template to handle user avatar changes. This permits users to upload a new image for their profile, which will be displayed immediately due to the page refreshing upon successful upload.