Change modal positioning from absolute to fixed
Updated modal positioning in 'CompanyModalRegister.vue', 'CompanyModalLogin.vue', 'CandidateModalRegister.vue' and 'CandidateModalLogin.vue' to be fixed instead of absolute. This change not only gives better control on modal positioning as it will now be relative to the browser window, but it also ensures smoother user experience across different screen sizes and scrolling behaviours.